requirements.txt
alembic==0.9.6
aniso8601==1.3.0
asn1crypto==0.23.0
certifi==2018.4.16
cffi==1.11.2
chardet==3.0.4
click==6.7
codecov==2.0.15
coverage==4.5.1
cryptography==2.3
fido2==0.3.0
flake8==3.3.0
flake8-isort==2.2.1
flake8-polyfill==1.0.1
Flask==0.12
Flask-Cors==3.0.2
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
gevent==1.3.5
greenlet==0.4.14
gunicorn==19.7.1
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
json-logging-py==0.2
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.3.2
py==1.4.34
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
PyJWT==1.6.4
pytest==3.2.3
pytest-cov==2.5.1
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
requests==2.19.1
six==1.11.0
SQLAlchemy==1.1.14
testfixtures==5.2.0
urllib3==1.23
Werkzeug==0.12.2
